+++ src/internal/from_files/js_parser.js
